% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_variables.R
\name{add_survey_variables}
\alias{add_survey_variables}
\title{Add derived variables to survey data}
\usage{
add_survey_variables(data, definitions)
}
\arguments{
\item{data}{Tibble with survey microdata.}

\item{definitions}{Named list of formulas defining new variables.}
}
\value{
Tibble with new variables added.
}
\description{
Add derived variables to survey data
}
