% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_design.R
\name{build_survey_design}
\alias{build_survey_design}
\title{Build a complex survey design}
\usage{
build_survey_design(
  data,
  weight,
  strata = NULL,
  cluster = NULL,
  fpc = NULL,
  nest = TRUE
)
}
\arguments{
\item{data}{Tibble with survey microdata}

\item{weight}{Weight variable name}

\item{strata}{Strata variable name (optional)}

\item{cluster}{Cluster (PSU) variable name (optional)}

\item{fpc}{Finite population correction variable (optional)}

\item{nest}{Logical. Whether PSUs are nested within strata}
}
\value{
A tbl_svy object
}
\description{
Build a complex survey design
}
