% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexSurvey_app.R
\name{ComplexSurvey_app}
\alias{ComplexSurvey_app}
\title{Run ShinyComplexSurvey application}
\usage{
ComplexSurvey_app()
}
\value{
Launches a Shiny application.
}
\description{
Launches the Shiny app bundled within the package.
}
