% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\name{generate_example_data}
\alias{generate_example_data}
\title{Generar datos simulados de encuesta compleja}
\usage{
generate_example_data(seed = 123)
}
\arguments{
\item{seed}{Semilla aleatoria}
}
\value{
tibble con datos simulados
}
\description{
Generar datos simulados de encuesta compleja
}
