% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_format.R
\name{plot_results_bar}
\alias{plot_results_bar}
\title{Plot survey estimates as bar chart}
\usage{
plot_results_bar(results)
}
\arguments{
\item{results}{Tibble returned by estimate_survey().}
}
\value{
A ggplot object.
}
\description{
Plot survey estimates as bar chart
}
