% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_design.R
\name{describe_survey_design}
\alias{describe_survey_design}
\title{Describe a survey design}
\usage{
describe_survey_design(design)
}
\arguments{
\item{design}{A tbl_svy object}
}
\value{
Tibble with basic design diagnostics
}
\description{
Describe a survey design
}
